+++ compound/src/test/kotlin/io/element/android/compound/screenshot/SemanticColorsTests.kt
@@ -14,7 +14,9 @@ import com.github.takahirom.roborazzi.captureRoboImage
 import io.element.android.compound.previews.ColorListPreview
 import io.element.android.compound.screenshot.utils.screenshotFile
 import io.element.android.compound.theme.ElementTheme
-import kotlinx.collections.immutable.toImmutableMap
+import io.element.android.compound.tokens.generated.compoundColorsHcDark
+import kotlinx.collections.immutable.ImmutableMap
+import kotlinx.collections.immutable.persistentMapOf
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.robolectric.annotation.Config
@@ -23,7 +25,7 @@ import org.robolectric.annotation.GraphicsMode
 @RunWith(AndroidJUnit4::class)
 @GraphicsMode(GraphicsMode.Mode.NATIVE)
 class SemanticColorsTests {
-    @Config(sdk = [34], qualifiers = "h2048dp-xhdpi")
+    @Config(sdk = [34], qualifiers = "h2500dp-xhdpi")
     @Test
     fun screenshots() {
         captureRoboImage(file = screenshotFile("Compound Semantic Colors - Light.png")) {
@@ -37,7 +39,27 @@ class SemanticColorsTests {
                         ColorListPreview(
                             backgroundColor = Color.White,
                             foregroundColor = Color.Black,
-                            colors = getSemanticColors().toMap().toImmutableMap()
+                            colors = getSemanticColors(),
+                        )
+                    }
+                }
+            }
+        }
+
+        captureRoboImage(file = screenshotFile("Compound Semantic Colors - Light HC.png")) {
+            ElementTheme(
+                compoundDark = compoundColorsHcDark,
+            ) {
+                Surface {
+                    Column(
+                        modifier = Modifier.padding(16.dp),
+                        verticalArrangement = Arrangement.spacedBy(10.dp)
+                    ) {
+                        Text("Compound Semantic Colors - Light HC")
+                        ColorListPreview(
+                            backgroundColor = Color.White,
+                            foregroundColor = Color.Black,
+                            colors = getSemanticColors()
                         )
                     }
                 }
@@ -55,7 +77,28 @@ class SemanticColorsTests {
                         ColorListPreview(
                             backgroundColor = Color.White,
                             foregroundColor = Color.Black,
-                            colors = getSemanticColors().toMap().toImmutableMap()
+                            colors = getSemanticColors(),
+                        )
+                    }
+                }
+            }
+        }
+
+        captureRoboImage(file = screenshotFile("Compound Semantic Colors - Dark HC.png")) {
+            ElementTheme(
+                darkTheme = true,
+                compoundDark = compoundColorsHcDark,
+            ) {
+                Surface {
+                    Column(
+                        modifier = Modifier.padding(16.dp),
+                        verticalArrangement = Arrangement.spacedBy(10.dp)
+                    ) {
+                        Text("Compound Semantic Colors - Dark HC")
+                        ColorListPreview(
+                            backgroundColor = Color.White,
+                            foregroundColor = Color.Black,
+                            colors = getSemanticColors()
                         )
                     }
                 }
@@ -64,9 +107,12 @@ class SemanticColorsTests {
     }
 
     @Composable
-    private fun getSemanticColors(): List<Pair<String, Color>> {
+    private fun getSemanticColors(): ImmutableMap<String, Color> {
         return with(ElementTheme.colors) {
-            listOf(
+            persistentMapOf(
+                "bgAccentHovered" to bgAccentHovered,
+                "bgAccentPressed" to bgAccentPressed,
+                "bgAccentRest" to bgAccentRest,
                 "bgActionPrimaryDisabled" to bgActionPrimaryDisabled,
                 "bgActionPrimaryHovered" to bgActionPrimaryHovered,
                 "bgActionPrimaryPressed" to bgActionPrimaryPressed,
@@ -80,6 +126,12 @@ class SemanticColorsTests {
                 "bgCriticalPrimary" to bgCriticalPrimary,
                 "bgCriticalSubtle" to bgCriticalSubtle,
                 "bgCriticalSubtleHovered" to bgCriticalSubtleHovered,
+                "bgDecorative1" to bgDecorative1,
+                "bgDecorative2" to bgDecorative2,
+                "bgDecorative3" to bgDecorative3,
+                "bgDecorative4" to bgDecorative4,
+                "bgDecorative5" to bgDecorative5,
+                "bgDecorative6" to bgDecorative6,
                 "bgInfoSubtle" to bgInfoSubtle,
                 "bgSubtlePrimary" to bgSubtlePrimary,
                 "bgSubtleSecondary" to bgSubtleSecondary,
@@ -94,6 +146,7 @@ class SemanticColorsTests {
                 "borderInteractivePrimary" to borderInteractivePrimary,
                 "borderInteractiveSecondary" to borderInteractiveSecondary,
                 "borderSuccessSubtle" to borderSuccessSubtle,
+                "iconAccentPrimary" to iconAccentPrimary,
                 "iconAccentTertiary" to iconAccentTertiary,
                 "iconCriticalPrimary" to iconCriticalPrimary,
                 "iconDisabled" to iconDisabled,
@@ -111,6 +164,12 @@ class SemanticColorsTests {
                 "textActionAccent" to textActionAccent,
                 "textActionPrimary" to textActionPrimary,
                 "textCriticalPrimary" to textCriticalPrimary,
+                "textDecorative1" to textDecorative1,
+                "textDecorative2" to textDecorative2,
+                "textDecorative3" to textDecorative3,
+                "textDecorative4" to textDecorative4,
+                "textDecorative5" to textDecorative5,
+                "textDecorative6" to textDecorative6,
                 "textDisabled" to textDisabled,
                 "textInfoPrimary" to textInfoPrimary,
                 "textLinkExternal" to textLinkExternal,
@@ -119,6 +178,7 @@ class SemanticColorsTests {
                 "textPrimary" to textPrimary,
                 "textSecondary" to textSecondary,
                 "textSuccessPrimary" to textSuccessPrimary,
+                "isLight" to if (isLight) Color.White else Color.Black,
             )
         }
     }